MA-14705 [Android] It will failed when format the sdcard as internal storage
It will have below error log when format:
 E vold    : private:179,66 failed to format: No such device.

The reproduce rate of this issue is very low, But we still need pick below
patch to make sure dm device is ready.

Wait for dm device to be ready before format
It can sometimes take a moment for the dm-device to appear after
creation, causing operations on it such as formatting to fail.
Ensure the device exists before create_crypto_blk_dev returns.

Change-Id: Ie1ab296ddcada3977497b683ac3faf0a05b91806
Signed-off-by: zhang sanshan <pete.zhang@nxp.com>