MA-15841-2 retry two times if open loop file fails
With the "LOOP_CTL_GET_FREE" ioctl command on the file descriptor of
"loop_control", a new loop device may be created if current loop device
not free, the new loop device may not be created after the
"LOOP_CTL_GET_FREE" command returns. add a retry if open the loop
devices failes.

This patch is used to fix failures in below two cts cases of the module
"CtsOsTestCases":
  android.os.storage.cts.StorageManagerTest#testMountAndUnmountTwoObbs
  android.os.storage.cts.StorageManagerTest#testMountAndUnmountObbNormal

Change-Id: I4867cf5c81481ebca602fbe61fe2bc675f8089aa
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>